+++ hack/generate-swagger-api.sh
@@ -23,8 +23,8 @@ swagger generate model -f api/swagger.yaml \
 
 swagger generate model -f api/swagger.yaml \
 	-t api -m types/image --skip-validator -C api/swagger-gen.yaml \
-	-n ImageDeleteResponseItem \
-	-n ImageSummary
+	-n ImageDeleteResponseItem
+#-n ImageSummary TODO: Restore when go-swagger is updated
 
 swagger generate model -f api/swagger.yaml \
 	-t api -m types/network --skip-validator -C api/swagger-gen.yaml \